Refactor: Extract executable content
Moves the content of the executable plugin script to a `process` method on `Twirp::ProtocPlugin` and documents current input/output parameters.

This allows us to keep the executable script extremely simple and gives us a "main"-style entry point with well documented input and output values that we can write specs against.