Consolidate service and client class template generation
Rather than using an additional indent_level, the generated output code is a little more obvious when we group the output together and manually indent the lines within the service and client class bodies.

We also in-line and delete the `rpc_line` helper in favor of keeping the output string content closer to the service class definition itself.