Add plumbing to allow for a `skip-empty` option.
This introduces the ability for caller to pass `--twirp_ruby_opt=skip-empty` to our plugin.

For now, we simply capture the flag if present on the `protoc` command line, and add some simple parameter validation. We also pass the command-line paramters down through to the `CodeGenerator` for further processing.

The logic for the flag itself remains unimplemented. That will be fleshed out in the next commit.