Add `dependency_proto_files` array to `FileDescriptorProto`
The `FileDescriptorProto` contains a `dependency` array of all of the imported file name strings, in the order in which they appear in the source file.

We want to create a parallel array, but instead of just the file names, we want the _references_ to those proto file descriptors. This enables future support for the proto file to map message types to their proper ruby modules, allowing the proto file descriptor to access the descriptors of dependencies all the way down.

We pair this new property with an override of the `CodeGeneratorRequest.decode` such that it automatically populates `dependency_proto_files` during creation.