Make `convert_to_ruby_type` aware of `ruby_package`.
Build an internal type map of all of the types in the current file, as well as in any imported files, respecting `ruby_package` along the way.

That trivializes the logic in `convert_to_ruby_type`; it turns into just a simple lookup from the type map and strips the module prefix if the type is in the same module.