Eliminate strange `rake spec` failures that `rspec` doesn't report.
Running `rspec` locally consistently shows `56 examples, 0 failures`. However, running `rake spec` locally occasionally shows `56 examples, 1 failure`. The failure is the same every time, and has to do with `dependency_proto_files` reporting unexpectedly as `nil`. Even re-running the failures with the same seed is intermittent (e.g. `--order rand:43823`)

Note that if I run the same command that `rake spec` outputs (to illustrate how it runs the specs), then I'm back to `56 examples, 0 failures`. There's something about the `rake` execution environment itself that seems to bring out the problem...

Nevertheless, I believe I have a solution: Instead of using the `proto_file.each` enumerator of `Google::Protobuf::RepeatedField`, we first convert the repeated field to an actual array, and then enumerate that. That's not quite enough though; _we further need to memoize the value_. The memoization is the key that makes the changes applied in the enumerator "stick" e.g. when accessing the `proto_file` repeated field in the future.