Fix standalone script load errors
When running the plugin locally as a script instead of a gem executable (via the `protoc` option `--plugin=protoc-gen-twirp_ruby=./exe/protoc-gen-twirp_ruby`), we need to explicitly add `lib` to the load path.

Note that the gem package handles this automatically via the gemspec `require_paths = ["lib"]`

This adds "lib" to the load path, but only if it's not already present.