Replace certain usages of `require_relative` with `require`
We switch to `require` for files that are meant to be loaded only a single time. This forces a consistent path to prevent accidental duplicate loads.

This fixes a runtime error when using the local plugin code to regenerate the example. It was  failing due to `require_relative` loading the `google/protobuf/descriptor_rb` twice (from different paths):

```
~/protoc-gen-twirp_ruby/lib/google/protobuf/descriptor_pb.rb:11:in `add_serialized_file': Unable to build file to DescriptorPool: duplicate file name google/protobuf/descriptor.proto (Google::Protobuf::TypeError)
  from ~/protoc-gen-twirp_ruby/lib/google/protobuf/descriptor_pb.rb:11:in `<top (required)>'
  from ~/protoc-gen-twirp_ruby/lib/twirp/protoc_plugin/descriptor_ext/file_descriptor_proto_ext.rb:4:in `require_relative'
```